+++ Mascoshop/Mascoshop/Mascoshop/src/main/java/com/Mascoshop/Repositorios/RepositorioProducto.java
@@ -9,13 +9,13 @@ import java.util.List;
 @Repository
 public interface RepositorioProducto extends JpaRepository<Producto, Integer> {
 
-    //Encontrar por nombre
-//    List<Producto> findByNombreContaining(String nombre);
-    //Econtrar por categoria
+
+    //Encontrar por categoria
     List<Producto> findByCategoriaProducto_IdCategoria(Integer categoriaId);
-    //Encontrar por animal :D
-    List<Producto> findByAnimal_IdAnimal(Long animalId);
+    //Encontrar por animal
+    List<Producto> findByAnimal_IdAnimal(Integer animalId);
     //Encontrar por marca
-    List<Producto> findByMarca_IdMarca(Long marcaId);
+    List<Producto> findByMarca_IdMarca(Integer marcaId);
 
 }
+
